Add files via upload
PREDICTING SURVIVORS IN THE TITANIC DATASET USING MACHINE LEARNING
To predict the survivors in the titanic ship using the dataset, we used the random forest classification machine learning model
Steps that were taken is explained below
First I imported the libraries I needed which are pandas, numpy, sklearn model selection for training testing and splitting data, sklearn ensemble for random forest classifier, sklearn metrics for accuracy score and classification report and imported warnings and filtered the warnings to ignore
The titanic dataset was  loaded and empty rows where dropped to clean the data and avoid inaccuracy or issues in prediction, rows that were empty and dropped included, deck, age, embarked. We changed the ‘alone’ column from Boolean to zeros and ones integers. I changed the gender name from female to 0 and male to 1 for the computer to understand. 
I created a new variable which is X and selected the columns that are imported for us to predict and that contain numerical value. 
I also created a new variable ‘y’  that includes the outcome we want to achieve. Then I checked in information of the X variables. After that was done, I went on to use the train-test-split to split the dataset into training dataset and testing dataset. I trained X variables and test X variables. I also trained ‘y’ variables and tested the ‘y’ variables. Then the test-size was indicated and picked a random-state. The test-size indicate the percentage of data that will be used for testing in decimals. I used 0.2 here which is 20% but the default value for the test-size is 0-25 which is 25%. The random-state used is 42. Any random state number can be used and it is to ensure reproducibility of the split.
I the created a RandomForestClassifier which contains the n-estimator and random-state can be inputted. RandomForestClassifier combines multiple decision trees. The n-estimator is the number of decision  trees to combine in the ensemble. It controls number of trees to build and combine to produce the final prediction. Our n-estimator here is 100 so it means the model will combine 100 decision trees to make predictions. I then Fittied the randomforestclassifier which was renamed as f-classifier to the training set (X-train, y-train). I also did the y-pred to test the prediction rate of the model
Then I evaluated the model by checking its accuracy, By comparing the predicted label y-predmwith the true table y-test and calculate the accuracy as the ratio of correctly classified samples to the total number of samples. The classification rep generated detailed classification report for the model predictions.
Classification-rep (classification report) provides a comprehensive summary of the model performance including:
1.	Precision:- which is the ratio of true positives to the sum of true positives and false positives. TP = TP + FP
2.	Recall:- ratio of true positions to the sum of true positives and false negatives. TP : TP + FN
3.	F1-score :- is the harmonic mean of precision and recall
4.	Supportnumber of actual samples in each class

Accuracy was outputted as 0.78 with this model